Updated SequenceWatchers and documentation.
Modified SequenceWatchers to use a HashMap instead of a Vec and the ability to
add and remove new sequences.

Fixed a bug in my test code. Having trouble filtering the parts of the stream
that are not supposed to have matching sequences.

Modified all documentation to use '///' instead of '/** */'. I liked the later
because it was easier to edit with my text editor, but it definitely made
navigating the code and finding things a lot harder. Lesson learned.