Minor improvement, documentation, and benchmarks
Removed the limitation on SequenceWatcher being able to handle empty sequences.
I decided that it wasn't good to have new panic on empty sequences to save
picoseconds in check by not having to check the size of the sequence.

Updated the documentation for the change, and coppied it to the README file.

Added benchmarks.